add ContentManager and cleanup Graphics class

+++ src/Castlevania/Animation/AnimationFrame.cpp
@@ -1,6 +1,6 @@
 #include "AnimationFrame.h"
 
-AnimationFrame::AnimationFrame(Sprite *sprite, int time)
+AnimationFrame::AnimationFrame(Sprite *sprite, unsigned long time)
 {
 	this->sprite = sprite;
 	this->time = time;
